# performance: add performance test