Goal was to measure TX/Sec as in https://github.com/neo-project/neo/issues/1321

In order to run it:
```
$ cd integration
$ go test -bench=. -benchmem
```

Result 1 (without singing TX):
```
10000            402421 ns/op          177370 B/op         90 allocs/op
PASS
ok      github.com/CityOfZion/neo-go/integration        4.360s
```
Meaning that we have 10 000 tx'es per 4.360s. It's 2294 TX / Sec.

Result 2 (with singing TX):
```
    4743            398582 ns/op          132878 B/op        467 allocs/op
PASS
ok      github.com/CityOfZion/neo-go/integration        2.530s
```
Meaning that we have 4743 tx'es per 2.530s. It's 1874 TX / Sec.

Next step would be to run performance test with e2e scenario.